processor/requirements.txt
Flask==2.2.2
scipy==1.12.0
numpy==1.23.5
mysql-connector-python==9.2.0
SQLAlchemy==2.0.36
nltk==3.8.1
scikit-learn==1.6.0
Pillow==9.2.0
PyPDF2==3.0.1
openpyxl==3.1.5
bs4==0.0.1
requests==2.32.3